requirements.txt
fastapi
uvicorn
pydantic
python-multipart
tortoise-orm
passlib
bcrypt
pyjwt